Extract "trigger", adapt to the new Matcher.
